backend/requirements.txt
# Innexar ERP - Django Multi-tenant SaaS
# Python 3.11+

# Django Core
Django==4.2.9
django-environ==0.11.2
psycopg2-binary==2.9.9
whitenoise==6.6.0

# Multi-tenancy
django-tenants==3.6.1

# API
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-spectacular==0.27.1

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-allauth==0.60.1

# Payments
stripe==4.2.0
dj-stripe==2.8.3

# Background Tasks
celery==5.3.6
redis==5.0.1
django-celery-beat==2.6.0
django-celery-results==2.5.1

# Email
django-anymail[resend]==10.3

# Storage
boto3==1.34.34
django-storages==1.14.2

# Internationalization
django-rosetta==0.10.0
django-modeltranslation==0.18.12
django-filter==24.3

# Utilities
python-dateutil==2.8.2
pytz==2024.1
pillow==10.2.0

# Development
django-debug-toolbar==4.3.0
django-extensions==3.2.3
ipython==8.21.0

# Production
gunicorn==21.2.0
sentry-sdk==1.40.0
